Filter the ingredients the user will see by its language.
Additionally, if the user has selected on his preference page that he wishes to
also see the ingredients in English (from the US Department of Agriculture), show
those too. This only makes sense if the user's language isn't English, as he will
be presented those in that case anyway, so also do a check for this.