First version of a more intelligent form handling in modal dialogs
If there is a form in the modal dialog (there usually is) prevent the
submit button from submitting it and do it with an AJAX request. If
there are errors (there is an element with the class 'ym-error' in
the result) reload the content back into the dialog so the user can
correct the entries. If there isn't assume all was saved correctly and
load that result into the page's main DIV (#main-content)