Allow Django to check the uniqueness of date-user pair in weight entries
This has to be like this because otherwise it's *very* difficult and
ugly to validate the uniqueness of unique_together field when one field
is excluded from the form (editable false). This does not pose any
security risk because the value from the form is ignored anyway and
the request's user always used.

--HG--
branch : mobile