Solved Issue #100 with PEP8 guidlines
I made an attempt to solve issue #100.
After the changes in code, the software seems to be working fine on my
machine without any errors.

Files changed: wger\excercises\views\equipment.py

In this view file, I created a list with translated equipment names,
sorted it and passed it to the template.

Followed PEP8.