Move tasks file to wger folder
This is really just needed so as to more easily make the packaged (pypi)
version work. The 'invoke' command (which is mapped to 'wger' when packaged or
installed via pip) needs a tasks.py file. Since the wger folder gets installed
as a python package, it is easier if we just move that file there as well.