Don't indent translatable text
Even if it doesn't look nice, it's better to not format the texts within
blocktrans, when the strings for translations are extracted, they are then
marked as new (because there are more spaces). Then, even if then transifex
offers a suggestion to use the old translation, this is a bit more work for the
translators.