Some changes to the installation scripts
These changes are necessary for the packaged version of the application
(pip install wger). Among other things and (hopefully!) to make the installation
more robust, all commands are routed through the 'wger' entry point, and calling
'invoke' on its own has been deprecated. Sadly, we had to revive an ugly hack
that changes the sys path, but hopefully this doesn't cause any problems.